Replaced old freqextr (slsclean) with new Python-based version
